Move rts.fs.mjs to platform specific bits
