+++ asterius/rts/browser/default.mjs
@@ -2,6 +2,59 @@
  * @file Implements browser-specific functionality.
  */
 
+class Device {
+  constructor(f, console_history) {
+    this.flush = f;
+    this.consoleHistory = console_history;
+    this.history = "";
+    this.buffer = "";
+    this.decoder = new TextDecoder("utf-8", { fatal: true });
+    Object.seal(this);
+  }
+
+  read() {
+    const r = this.history;
+    this.history = "";
+    return r;
+  }
+
+  write(buf) {
+    const str =
+      typeof buf === "string"
+        ? buf
+        : this.decoder.decode(buf, { stream: true });
+    if (this.consoleHistory) {
+      this.history += str;
+    }
+    this.buffer += str;
+    const segs = this.buffer.split("\n");
+    this.buffer = segs.pop();
+    for (const seg of segs) {
+      this.flush(seg);
+    }
+    return buf.length;
+  }
+}
+
+class MemoryFileSystem {
+  constructor(console_history) {
+    this.files = [
+      undefined,
+      new Device(console.log, console_history),
+      new Device(console.error, console_history)
+    ];
+    Object.freeze(this);
+  }
+
+  readSync(fd) {
+    return this.files[fd].read();
+  }
+
+  writeSync(fd, buf) {
+    return this.files[fd].write(buf);
+  }
+}
+
 class Posix {
   constructor(memory, rtsConstants) {
     this.memory = memory;
@@ -87,6 +140,7 @@ export default {
      */
     resolution: 1000000
   },
+  fs: MemoryFileSystem,
   posix: Posix
 };
